// src/types/stockItem.type.ts
export interface StockItem {
    id: number;
    category: string;
    itemType: string;
    itemCode: string;
    name: string;
    description: string;
    quantity: number;
    uom: string;
    itemGroup: string;
    itemCategory: string;
    element: string;
    subElement: string;
    subSubElement: string;
    isActive: boolean;
}